Add Schema/Data Creation
After doing some research, I found that with MySQL  is not as easy to use like the midterm since the midterm was using an in-memory db. It is possible to have a programmatically defined user but for now you need to set the login in
C:\Users\Aidan\Documents\GitHub\final-project\final-project\src\main\resources\application.properties
I would recommend just making a new user that matches what is already set, but if you do change it, make sure not to commit that change.

You do not need to create the database, however.

Also, I have made a test case that passes and a test case that fails just to check the JUnit.